langgraph/pyproject.toml
[project]
name = "langgraph-TIL"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-tavily>=0.1.6",
    "langchain[openai]>=0.3.25",
    "langgraph>=0.4.3",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.10",
]
